Organize repo and add simple home screen
