--- .expo/devices.json
@@ -1,3 +1,0 @@
-{
-  "devices": []
-}
