--- package.json
@@ -1,18 +1,0 @@
-{
-  "dependencies": {
-    "@expo/metro-runtime": "~5.0.4",
-    "@react-navigation/native": "^7.1.10",
-    "@react-navigation/native-stack": "^7.3.14",
-    "expo": "^53.0.11",
-    "react-dom": "19.0.0",
-    "react-native-gesture-handler": "~2.24.0",
-    "react-native-reanimated": "~3.17.4",
-    "react-native-safe-area-context": "5.4.0",
-    "react-native-screens": "~4.11.1",
-    "react-native-web": "^0.20.0"
-  },
-  "devDependencies": {
-    "@types/react": "~19.0.10",
-    "typescript": "~5.8.3"
-  }
-}
